# esdebitami_retake_ai/src/main/resources/public/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vas-Ai",
    "description": "API del modulo Vas-Ai\n<br>Timestamp generazione: 2020-02-16T17:11:52.00",
    "version": "2.3.0"
  },
  "paths": {
    "/vas-ai/nlp/stemming": {
      "post": {
        "tags": [
          "Lemming"
        ],
        "summary": "Restituisce la frase fornita in ingresso stemmata",
        "description": "Restituisce la frase fornita in ingresso stemmata",
        "operationId": "Stemming",
        "parameters": [
          {
            "name": "sentence",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Frase stemmata correttamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "items": {
                    "$ref": "VAS-Response/content"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/vas-ai/nlp/lemming": {
      "post": {
        "tags": [
          "Stemming"
        ],
        "summary": "Genera il lemma della frase in ingresso e lo fornisce in output",
        "description": "Genera il lemma della frase in ingresso e lo fornisce in output.",
        "operationId": "Lemming",
        "parameters": [
          {
            "name": "sentence",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Il lemma della frase fornita",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "VAS-Response/content"
                }
              }
            }
          }
        }
      }
    },
    "/vas-ai/nlp/landetect": {
      "post": {
        "tags": [
          "LanguageDetection"
        ],
        "summary": "Identifica la lingua utilizzata analizzando la frase in ingresso.",
        "description": "Identifica la lingua utilizzata analizzando la frase in ingresso.",
        "operationId": "landetect",
        "parameters": [
          {
            "name": "sentence",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Le due lettere identificanti la lingua utilizzata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          }
        }
      }
    },
    "/vas-ai/nlp/generateLink": {
      "post": {
        "tags": [
          "Link generation"
        ],
        "summary": "Consente la generazione automatica di link sulla base del sito e delle keyword inserite.",
        "description": "Consente la generazione automatica di link sulla base del sito e delle keyword inserite.",
        "operationId": "generateLink",
        "parameters": [
          {
            "name": "siteName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "researchKey",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Gli n link corrispondenti ai parametri di ricerca",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": ""
                }
              }
            }
          }
        }
      }
    },
    "/vas-ai/nlp/findTemplate": {
      "post": {
        "tags": [
          "Identificazione template specifico"
        ],
        "summary": "Cerca il template specifico tale da soddisfare le richieste dell'utente basandosi sulla frase inserita in input.",
        "description": "Cerca il template specifico tale da soddisfare le richieste dell'utente basandosi sulla frase fornita in input.",
        "operationId": "findTemplate",
        "parameters": [
          {
            "name": "sentence",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200 - singolo match": {
            "description": "L'id tale da soddisfare la richiesta dell'utente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "VAS-Response/content"
                }
              }
            }
          }
        },     
          "200 - nessun match": {
            "description": " Non è stato trovato alcun match con la frase in ingresso, il campo templateId è = -1 , il campo ambiguity è = non è stato trovato alcun match ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "VAS-Response/content"
                }
              }
            }
          },
          
          "200 - match multiplo": {
            "description": " la frase inserita matcha con più template, il campo templateId è = -1, il campo ambiguity è = 'potresti essere più specifico? sono stati trovati N template con una corrispondenza' '",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "VAS-Response/content"
                }
              }
            }
          }
        
        
        
        
      }
    },
    "/vas-ai/nlp/findNextQuestionExp": {
      "post": {
        "tags": [
          "Identificazione domanda successiva"
        ],
        "summary": "Cerca l'id della domanda successiva, basandosi sulla frase inserita dall'utente. Effettua una gestione di casi di disambiguità e/o di elevato numero di corrispondenze",
        "description": "cerca l'id della domanda successiva, basandosi sulla frase inserita dall'utente. La gestione del risultato avviene attraverso due steps successivi: [1] si cerca un uguaglianza tra la frase inserita e le keywords delle domande successive [2] si controlla se la frase in ingresso contiene la n-esima keyword o se il token della frase è uguale al token della n-esima keyword",
        "operationId": "findNextQuestionExp",
        "parameters": [
          {
            "name": "sentence",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "templateId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "questionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200 - uguaglianza frase keyword": {
            "description": "L'id della domanda successiva, l'info sullo stato di foglia della domanda successiva. La risposta è visualizzata solamente se la frase inserita uguaglia una delle n keywords",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "Vas-Response/content"
                	  }
                  }
            	}
          	}
          },
          "200 - soluzione unica": {
            "description": "L'id della domanda successiva, l'info sullo stato di foglia della domanda successiva. La risposta è visualizzata se il match non è unico come nel caso precedente, ma tutti i match riconducono ad una sola domanda successiva",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "Vas-Response/content"
                  }
                }
              }
            }
          },
          "200 - disambiguazione": {
            "description": "Si verifica se le corrispondenze frase-key consentono di identificare 2 possibili domande successive. In tutti i casi successivi la risposta è del tipo nextQuestionId=-1, foglia=false, ambiguity=x dove x varia a seconda del caso. In particolare : [1] il numero di corrispondenze è pari a 2, ambiguity= forse cercavi + le due keyword con corrispondenza, [2] il numero di corrispondenze è >2, ambiguity =  potresti essere più specifico",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "Vas-Response/content"
                  }
                }
              }
            }
          },
         "200 - numero eccessivo di match": {
          "description": "Si verifica se le corrispondenze frase-key consentono di identificare più di 2 possibili domande successive. La risposta è del tipo nextQuestionId=-1, foglia=false, ambiguity =  potresti essere più specifico",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "Vas-Response/content"
                  }
                }
              }
            }
          },
          "200 - nessun match": {
          "description": "Si verifica se le corrispondenze frase-key non consentono di identificare nessuna domanda successiva. La risposta è del tipo nextQuestionId=-1, foglia = false, ambiguity = non è stata trovata alcuna corrispondenza",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "Vas-Response/content"
                  }
                }
              }
            }
          }
          
          
          
          
          
        }
      }
    }
  }
}